DocAboutComplete

+++ frontend/src/utils/formateDate.js
@@ -1,5 +1,5 @@
 export const formateDate =(date, config)=>{
-    const defaultOptions ={day:'numeric',month:'long',year:'numeric'}
+    const defaultOptions ={day:'numeric',month:'short',year:'numeric'}
     const options = config ? config :defaultOptions
     return new Date(date).toLocaleDateString('en-Us',options)
